Remove recent peer dependency bumps
In a previous release, some packages received new versions, and
corresponding entries in `dependencies`, `devDependencies`, and
`peerDependencies` were bumped to match. However, for packages that only
received minor or patch bumps, it was not necessary to bump the peer
dependencies. (These bumps may have been required by a rule that either
the release tool or the `core` repo enforces.) Reverting these changes
avoids having to create major releases for the packages in which these
peer dependencies were bumped.