assets-controllers: Remove AllowedActions, AllowedEvents
These types were mistakenly exported in a previous commit when
converting NftDetectionController to BaseController v2. They can still
be exports in `NftDetectionController.ts`, but just not be exports from
the perspective of the whole `assets-controllers` package. This commit
takes these exports away by replacing the `*` export in `index.ts` with
explicit names and ensures not to include the two aforementioned types.