fix: TransactionController afterSign hook should be allowed to modify the transaction (#4343)
> [!NOTE]
> This change is intended to be cherry-picked into MetaMask/extension
v11.17.0 via an upgrade to the @metamask/transaction-controller package.
## Explanation
**Issue:**
A recent update to the transaction-controller has made the
TransactionMeta object passed to the `afterSign` hook frozen. This
change prevents adding new properties, leading to the error: “Cannot add
property custodyId, object is not extensible.” This bug is breaking all
transactions for MMI as the original txMeta cannot store required
properties like custodyId.

**Fix:**
We deep clone the transaction meta before passing it to the hook. A deep
clone is used because transactionMeta is recursively frozen by immer.
This fix was intended to minimize the change for the cherry-pick going
into v11.17.0. A longer term solution might involve using immer more
throughout the TransactionController.ts file to make it clearer when a
transactionMeta is being mutated and how.

**Testing:**
This fix was applied to and verified with the MMI extension.

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/transaction-controller`

- **FIXED**: afterSign hook is now able to modify the transaction

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate