Release 158.0.0 (#4351)
This is a redo of commit c2e675256a5dc7f5b73e23f58867e8f29098f689, which
did not result in a release because it did not update the version of the
root package. This commit fixes that.

The remainder of this description was copied from that commit:

The goal of this PR is to release any changes in packages that have not
yet been released. This is a prerequisite to upgrading all packages to
Node 18. Hence, this PR includes releases for most packages.

See updates to changelogs for more.

---

Co-authored-by: Jongsun Suh <jongsun.suh@icloud.com>
Co-authored-by: Monte Lai <monte.lai@consensys.net>
Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>
Co-authored-by: Derek Brans <dbrans@gmail.com>
Co-authored-by: Brian Bergeron <brian.e.bergeron@gmail.com>