Release 159.0.0
The purpose of this PR is primarily to upgrade all packages to a minimum
version of Node 18.18.

While preparing this release I noticed that the previous release did not
create a new version for `@metamask/controller-utils` because the
version was not incremented when it should have been. So that version
has been properly incremented and the changelog notes that were added
in this previous release have been moved to the right place.