feat: Allow overwriting built-in keyring builders
The KeyringController comes with two built-in keyrings: Simple and HD.
Unfortunately it's impossible to overwrite these because when we build
a new keyring, we look for the first keyring builder in the list that
matches the type we want to build, and the built-in keyrings are always
first.

The order has been switched so that built-in keyrings come second,
after custom keyring builders. This allows them to be overwritten.

This makes it possible to test behavior that is specific to simple or
HD keyrings. This is something that I wanted to do in the mobile
repository.