feat: Allow overwriting built-in keyring builders (#4362)
## Explanation

The KeyringController comes with two built-in keyrings: Simple and HD.
Unfortunately it's impossible to overwrite these because when we build a
new keyring, we look for the first keyring builder in the list that
matches the type we want to build, and the built-in keyrings are always
first.

The order has been switched so that built-in keyrings come second, after
custom keyring builders. This allows them to be overwritten.

This makes it possible to test behavior that is specific to simple or HD
keyrings. This is something that I wanted to do in the mobile
repository.

## References

N/A

## Changelog

### `@metamask/keyring-controller`

#### Added

- Add support for overwriting built-in keyring builders for the Simple
and HD keyring.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate