AccountTrackerController: Remove ref to providerConfig (#4306)
The `providerConfig` state property is being removed from
NetworkController. Currently this property is used in
AccountTrackerController to access the type of the `chainId` property,
but we know that this type is `Hex`, so we can simply use it directly.
This commit makes that change so that we can fully drop
`providerConfig`.