refactor: update account tracker to use selected account (#4225)
## Explanation

This PR updates the `AccountTrackerController` to use `selectedAccount`
and InternalAccounts. The callbacks of the preferences controller are
removed and replaced by actions to the controller messenger. The
subscription of `onPreferencesStateChange` has been replaced with a
subscription to the controller messenger.

## References

Fixes https://github.com/MetaMask/accounts-planning/issues/381

## Changelog

### `@metamask/assets-controllers`

- **CHANGED**: Removed `getIdentities`, `getSelectedAddress` and
`onPreferencesStateChange` from the constructor arguments of the
`AccountTrackerController`
- **BREAKING**: Require a controller messenger in the constructor of the
`AccountTrackerController`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>