Remove providerConfig from NetworkController
Historically, the `providerConfig` property in NetworkController has
been used to track the currently selected network as well as provide
access to information about that network. The selected network is now
tracked via `selectedNetworkClientId`, and information about that
network can be retrieved by looking at the `networkConfigurations`
property or the `configuration` property on the NetworkClient interface.
This means that we no longer need `providerConfig` and we can remove
this redundant state.