upgrade TokenRatesController to BaseControllerV2 (#4314)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes.

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

In this, the TokenRatesController has been upgraded to BaseControllerV2.
The upgrade includes TokenRatesController inheriting BaseController v2
instead of BaseController v1. This affects the constructor and the also
the way state is updated inside the controller, but it also prompts
other changes:

This also includes the changes to the unit tests, so that all the tests
uses `withController` pattern.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->
Fixes #4076 

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

`@metamask/assets-controller`
### Added

- New types for `TokenRatesController `messenger actions 
  - `TokenRatesControllerGetStateAction`
- New types for `TokenRatesController` messenger events
  - `TokenRatesControllerStateChangeEvent`

### Changed

- **BREAKING:** Changed superclass of `TokenRatesController` from
BaseController v1 to BaseController v2
- **BREAKING:** Renamed `TokenRatesState` to `TokenRatesControllerState`

### Removed

- **BREAKING:** Removed `TokenRatesConfig` type

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate