Restore ESLint warnings as errors (ignoring them for now) (#4382)
Currently, `yarn lint` produces a bunch of warnings from ESLint. These
warnings were added in a previous commit that upgraded ESLint packages
as a way to avoid fixing lint violations created by the upgrade.
However, this causes two problems:

1. They produce a lot of noise that makes it very difficult to find true
errors, especially when looking at a CI run.
2. They allow new instances of these violations to show up (because
warnings don't cause `eslint` to fail).

This commit removes the overrides from the ESLint config for these rules
so that they go back to being errors and adds `eslint-disable`
directives above the lines that cause the violations. Note that this is
not intended to be a long-term solution, and in the future, we should
dedicate time to either fixing the violations or explaining why ignoring
them is necessary.