Clean build artifacts locally before linting
If you run `yarn build` and then run ESLint (either in your editor or
the `eslint` CLI), you may see false positives — i.e., you may see lint
errors show up on CI, but when you try to reproduce them locally, they
mysteriously disappear. I am not 100% sure why this happens, but I
believe it's because when the ESLint rules that make use of TypeScript
run, TypeScript will parse the type declaration files that were
generated via `yarn build` and use them as a reference instead of using
the live version of the TypeScript files, so if the type declaration
files are out of date, then the results of ESLint will also be out of
date. Ensuring that all build artifacts are removed before linting makes
these false positives go away.

This commit adds a clean step to both `yarn lint` and `yarn lint:fix`,
and it also ensures that the clean step precedes the build step for
consistency.