Remove tests for scripts/create-package/index.ts
The `create-package` utility is used as a quick way to make a new
monorepo package. The entrypoint, or executable, for this utility is
stored at `scripts/create-package/index.ts`. There are tests for this
file, which is good. However, running them may cause Jest to exit with a
non-zero code, even if they pass. And because `yarn test` runs in silent
mode by default, `yarn test` will then show no output, which can be
quite confusing.

Jest could be exiting in this way because by its nature, `index.ts`
begins running the utility as soon as it's loaded, but this file is
asynchronous and the test loads it without awaiting. So it's possible
Jest is exiting with non-zero because it detected an asynchronous action
that was never resolved.

Whatever the reason for failure, we don't need to test `index.ts`,
because the majority of its behavior is covered in the tests for
`cli.ts`. So, this commit simply removes the tests for `index.ts`.