Feat/integrate notifications controllers (#4416)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This adds some core controllers for the notification services that are
consumed in both mobile and extension. (Maintained by the notifications
team).

Controllers are:
- `ProfileSyncController` - which contains 2 controllers
(`AuthenticationController` and `UserStorageController`)
- These are used for adding or consuming authenticated endpoints and
also make use of the profile syncing feature across platforms and
devices.

- `NotificationServicesController` - which also contains 2 controllers
(`NotificationServicesController` and
`NotificationsServicesPushController`)
- Both of these controllers are tightly coupled. The first manages pull
based notifications and creation of resource; and the second is
responsible for push notifications.

NOTE - these controllers should be treated as V0.x.x, as they are under
development on extension and mobile (iron out any issues).

Previous PR: https://github.com/MetaMask/core/pull/4320
(closed this as it was getting to unruly to work in.)

## References

Here is a loom recording for the new controllers.

https://www.loom.com/share/4e95a8fcc2ae4d81b737265fc75571c5?sid=8247a832-3f2e-4837-ab74-30c45524ccfe

Future Improvements (as separate PRs)
1. Lets tidy up and improve Push Notifications for Mobile (after
dog-fooding) - the implementation may differ on mobile.
2. I also want to improve imports so we can do paths (e.g.
`@metamask/profile-sync-controller/sdk`) instead of the global named
exports have.
3. Decoupling this from extension has shown some annoying messaging
system actions back and forth (leading to some circular dependencies due
to importing types). This has been resolved for now, but I want our team
to rethink how we are using the messaging system to be more streamlined.
- I don't want UserStorage controller calling actions to Notifications
(we should not do this)
- I don't really like how tied the notifications and push notifications
do communication back and forth. I would rather we orchestrate
communication in 1 controller (so communication is 1 way).

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/profile-sync-controller`

- **ADDED**: new `AuthenticationController`
- **ADDED**: new `UserStorageController`

### `@metamask/notfication-services-controller`

- **ADDED**: new `NotificationServicesController`
- **ADDED**: new `NotificationServicesPushController`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: JSoufer <jonathan.ferreira@consensys.net>
Co-authored-by: Jonathan Ferreira <44679989+Jonathansoufer@users.noreply.github.com>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>