feat: change `queued-request-controller` methods array params to callbacks (#4423)
## Explanation

The existing `methodsWithConfirmation` array param is not sufficient for
determining if a method will cause a confirmation to be presented to the
user or not because some methods such as `eth_requestAccounts` may only
cause a prompt in some scenarios but not others.

This PR replaces the array params used in the QueuedRequestController
and its middleware to be callbacks that accept the entire request object
instead.

## References

See: https://github.com/MetaMask/metamask-extension/pull/25310

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/queued-request-controller`

- **BREAKING**: `QueuedRequestController` constructor no longer accepts
the `methodsRequiringNetworkSwitch` array param. It's now replaced with
the `shouldRequestSwitchNetwork` function param which should return true
when a request requires the globally selected network to match that of
the dapp.
- **BREAKING**: `createQueuedRequestMiddleware` no longer accepts the
`methodsWithConfirmation` array param. It's now replaced with the
`shouldEnqueueRequest` function param which should return true when a
request should be handled by the QueuedRequestController


## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate