Enable `resetMocks` Jest configuration option (#4417)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

Jest has a configuration option `resetMocks`. Equivalent to calling
`jest.resetAllMocks()` before each test, it helps to ensure that tests
are run in isolation by removing fake implementations for each
registered mock function, thereby resetting its state.

This option is enabled by default for new Jest projects but has been
disabled in this repo for a very long time. For some test suites,
`jest.resetAllMocks()` (or some variant) has been added to a
`beforeEach` or `afterEach` to simulate this option, but overall this is
not the case, and some tests were written which assumed that mock
functions were not being reset between tests.

This commit enables the aforementioned configuration option, fixes tests
that fail as a result of this, and removes manual calls to
`jest.resetAllMocks()` (or the like).

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

Fixes #745.

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

(N/A)

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: jiexi <jiexiluan@gmail.com>
Co-authored-by: Alex Donesky <adonesky@gmail.com>
Co-authored-by: Prithpal Sooriya <prithpal.sooriya@consensys.net>
Co-authored-by: Shane <jonas.shane@gmail.com>
Co-authored-by: Derek Brans <dbrans@gmail.com>
Co-authored-by: Monte Lai <monte.lai@consensys.net>