fix: Fix `create-package` tests that cause test command to fail
The test for `create-package` was causing the test command to fail with
a non-zero exit code even when tests passed. This was because the code
under test was setting `process.exitCode`, and this causes Jest to exit
with that same code.

The test was updated to mock `process` to ensure that the real
`process.exitCode` was not changed. Additionally, an `afterEach` hook
has been added for all script tests that will catch such issues in the
future.