fix: Fix `create-package` changes to `tsconfig.build.json`
The `create-package` tool automatically updates `tsconfig.build.json`
with an entry for the new package, but this entry is invalid. It should
reference the package-level `tsconfig.build.json` file, but instead it
just references the package directory (so the `tsconfig.json` file is
used instead.

This mistake resulted in very confusing build errors when I recently
created a new package.

This has been corrected, and tests have been updated accordingly.