fix: Fix `update-readme-content` script
The `update-readme-content` script will automatically update the
package graph in the README. This script was omitting the first
package in the list, the `accounts-controller` package.

The script has been updated to stop removing this package, and the
script was run to update the current README (one dependency between
core packages was missing).

It's unclear why the first line was removed in the first place. It's
possible that an earlier version of Yarn had one additional line in the
output.