Ensure functions that should return Hex, do (#193)
Some functions that return hex strings satisfying the `Hex` type return
a `string` type instead. This makes it slightly inconvenient to use
these functions in conjunction with others, and it is common in practice
to use a type assertion as a workaround. This commit rectifies this
issue by assigning these functions proper return types:

- `getChecksumAddress`
- `numberToHex`
- `bigIntToHex`

As such this change is **breaking**.