feat: make SafeEventEmitterProvider compatible with eth req libraries (#4422)
## Explanation

- Aligning SafeEventEmitterProvider to EIP-1193 by adding a `request`
method and deprecating `sendAsync` (which aren't part of the spec)
- Ensuring that `request` does not require `id` and `jsonrpc` to be set,
filling them in if they are missing

## References

Fixes #4095 

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/eth-json-rpc-provider`

- **ADDED**: request method type compatible with ethereum request
libraries.
- **CHANGED**: markes sendAsync as `deprecated` for usage in favor of
`request`.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate