chore(keyring-controller): Improve `createNewVaultAndKeychain` tests and docs
The tests for `createNewVaultAndKeychain` have been improved. Many of
them did not call the function-under-test at all, or they called it in
the wrong circumstances.

The inline comment for the function was updated to be more accurate. It
implied that an existing vault would be wiped out, but that's not
necessarily true.