chore(keyring-controller): Improve `createNewVaultAndKeychain` tests and docs (#4520)
## Explanation

The tests for `createNewVaultAndKeychain` have been improved. Many of
them did not call the function-under-test at all, or they called it in
the wrong circumstances.

The inline comment for the function was updated to be more accurate. It
implied that an existing vault would be wiped out, but that's not
necessarily true.


## References

N/A

## Changelog

None

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate