Fix permission-controller lint violations
There are some `permission-controller` tests which attempt to push an
async middleware function onto an engine. Although this works in
practice, `@typescript-eslint` balks at this, because the type for
`JsonRpcEngine.push` says it takes a `JsonRpcMiddleware`, and a
`JsonRpcMiddleware` has a return type of `void`, not `Promise<void>`. We
could change that return type to `void | Promise<void>` but that would
cause other changes we may not want to make. So this commit merely
ignores the lint violations.

Interestingly, the lint violations caused by this discrepancy only
appear locally and not on CI. I am not sure why this is.