fix: notification services controller missing env vars (#4530)
## Explanation

If platforms try to use this controller but pass invalid environment
variables, we will make unnecessary calls to fetch feature announcements
that will fail. (This mostly happens in Extension, where some
controllers are in JS and do not provide type errors).

NOTE - that a new service is being developed to deliver feature
announcements, which might allow us to not require platforms to pass in
an environment.

## References

[NOTIFY-848](https://consensyssoftware.atlassian.net/browse/NOTIFY-848)

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/notification-services-controller`

- **CHANGED**: Updated notification mocks to use realistic tx hashes
- **FIXED**: Logic to check feature announcement environment before
fetching

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate