refactor: add login response validation (#4541)
## Explanation

Profile Syncing. Adds login response validation from the developer/users
storage medium.
The user could lie and pass in an invalid login response, so this makes
sure that there is some validation around the response shape.

NOTE - we can improve this in the future by adding a runtime
parser/validator like Zod. But this is out of scope and we can add in a
separate PR.

## References

[NOTIFY-881](https://consensyssoftware.atlassian.net/browse/NOTIFY-881)

## Changelog

### `@metamask/profile-sync-controller`

- **ADDED**: Validation around the LoginResponse shape.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate