fix: Compress transaction history upon update
The TransactionController has been updated to compress transaction history if
it exceeds the max transaction history size, which for now has been set to 100.
Each time a new entry is added to a transaction history already at max size, we
merge two entries to make room for the new one.

Note that we never compress entries used for display in the transaction
activity log, because compressing those entries might hide events that we want
to display. If there are no non-displayed entries to compress, history is
allowed to exceed the max size.

This is a temporary solution to prevent unbounded growth of the transaction
history. While technically it is still unbounded at this level because we don't
strictly limit displayed history entries, we don't know of any cases where
displayed entries can be repeated a significant number of times, so this will
solve that problem in practice.

Fixes #4549