chore: Remove redundant dependencies
In various packages, we have a package listed as both a peer dependency and as
a dependency. This was done by mistake, the dependency entries aren't actually
used in practice, as we require the package to be installed as a "peer"
package anyway. It's also incorrect to list other controllers as a dependency
in this case because we don't know which version we need. In these cases we
need it to match the version setup with the global messenger, which we don't
know.