Convert constraints from Prolog to JavaScript (#4546)
The existing Yarn constraints have been difficult to maintain because
they are written in Prolog. More recent versions of Yarn support
JavaScript-based constraints, so this commit rewrites the existing
constraints in JavaScript.

Instead of `constraints.pro`, all constraints are now kept in
`yarn.config.cjs`. This file can be used for other things in the future
besides constraints. (And constraints can be used in the future to check
more than just dependencies!)

Note that we have had discussions in the past around the peer dependency
constraints, and we know that they are not quite correct. This will be
fixed in another PR.

---------

Co-authored-by: cryptodev-2s <109512101+cryptodev-2s@users.noreply.github.com>