feat: update notifications controllers from extension fixes. (#4573)
## Explanation

There were a set of additional extension fixes that also need to be
ported over to the shared libraries.
We are actively working on replacing the extension controllers with the
shared libraries (this way we don't need to do this weird back and forth
syncing).

## References

N/A

## Changelog

### `@metamask/notification-services-controller`

- **ADDED**: New events when notifications list is updated or when
notifications are marked as read.
- **ADDED**: shared function `toRawOnChainNotification` that is used to
polyfill and complete raw on chain notifications (before they get
processed).
- **FIXED**: update push notifications to correctly use
polyfilled/complete raw notifications.
- **ADDED**: new type `UnprocessedOnChainRawNotification` to signify the
actual raw data we receive from servers.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate